Improved mode() method
The previous code failed and reported 'w' for files that had been read entirely
(<$fh> returned undef). Also, it did not save the mode() determined from non-
filehandles, '?'.

The new approach to determine if a stream is readable or writable is to read or
write 0 characters to or from it and check return status to find out if that
was successful.

+ some general code tabulation fixes, cleanups, POD clarifications